+++ src/test/java/racingcar/domain/CarTest.java
@@ -1,0 +1,36 @@
+package racingcar.domain;
+
+import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
+
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
+
+class CarTest {
+
+    @DisplayName("Test for a valid name length")
+    @ParameterizedTest
+    @ValueSource(strings = {"1", "12", "123", "1234", "12345"})
+    void checkLengthTest(String name) {
+        assertThatCode(() -> Car.createByName(name))
+            .doesNotThrowAnyException();
+    }
+
+    @DisplayName("Exception test for null name")
+    @Test
+    void checkNullExceptionTest() {
+        Assertions.assertThatThrownBy(() -> Car.createByName(null))
+            .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("Exception test for over name length limit")
+    @ParameterizedTest
+    @ValueSource(strings = {"", "    ", "123456"})
+    void checkLengthExceptionTest(String name) {
+        Assertions.assertThatThrownBy(() -> {
+            Car.createByName(name);
+        }).isInstanceOf(IllegalArgumentException.class);
+    }
+}
